Added: program to find max in 3 Floats

+++ src/main/java/com/bridgelabz/maximum/Maximum.java
@@ -4,8 +4,8 @@ package com.bridgelabz.maximum;
  * Hello world!
  */
 public class Maximum {
-    public static Integer getmaximum(Integer value1, Integer value2, Integer value3) {
-        Integer max = value1;
+    public static Float getmaximum(Float value1, Float value2, Float value3) {
+        Float max = value1;
         if (value2.compareTo(value1) > 0 && value2.compareTo(value3) > 0)
             max = value2;
         else if (value3.compareTo(value2) > 0 && value3.compareTo(value1) > 0)
